+++ BoplTranslator/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v2.0.0
+New API for making custom translatable texts
+Better guid based tranlsation system
+
 # v1.1.0
 Updated to game version 2.2.0